pyproject.toml
[project]
name = "enterprise-ai-powered-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas==2.2.3",
    "plotly==5.24.1",
    "python-dotenv==1.0.1",
    "streamlit==1.39.0",
    "supabase==2.10.0",
]

[tool.uv.workspace]
members = [
    "data-gen",
]
